+++ helpdesk/templates/helpdesk/ticket_desc_table.html
@@ -221,7 +221,7 @@ copy_public_link = function() {
                             <ul>
                             {% for followup in ticket.followup_set.all %}
                             {% for attachment in followup.followupattachment_set.all %}
-                            <li><a href='{{ attachment.download_url }}'>{{ attachment.filename }}</a> ({{ attachment.mime_type }}, {{ attachment.size|filesizeformat }})
+                            <li><a href='{{ attachment.donwload_attachment }}'>{{ attachment.filename }}</a> ({{ attachment.mime_type }}, {{ attachment.size|filesizeformat }})
                             {% if followup.user and request.user == followup.user %}
                             <a href='{% url 'helpdesk:attachment_del' ticket.id attachment.id %}'><button class="btn btn-danger btn-sm"><i class="fas fa-trash"></i></button></a>
                             {% endif %}
